web development/vn tech internship platform/vnfinal/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
colorama==0.4.6
    # via
    #   -r requirements.in
    #   qrcode
contourpy==1.3.0
    # via
    #   -r requirements.in
    #   matplotlib
cycler==0.12.1
    # via
    #   -r requirements.in
    #   matplotlib
dataclasses==0.6
    # via -r requirements.in
django==5.1.2
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-autocomplete-light
    #   django-cors-headers
    #   django-qr-code
    #   django-select2
    #   djangorestframework
django-appconf==1.0.6
    # via
    #   -r requirements.in
    #   django-select2
django-autocomplete-light==3.11.0
    # via -r requirements.in
django-cors-headers==4.5.0
    # via -r requirements.in
django-qr-code==4.1.0
    # via -r requirements.in
django-select2==8.2.1
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   pymongo
fonttools==4.54.1
    # via
    #   -r requirements.in
    #   matplotlib
kiwisolver==1.4.7
    # via
    #   -r requirements.in
    #   matplotlib
matplotlib==3.9.2
    # via -r requirements.in
mongoengine==0.29.1
    # via -r requirements.in
mysqlclient==2.2.5
    # via -r requirements.in
numpy==2.1.2
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==24.1
    # via
    #   -r requirements.in
    #   matplotlib
pillow==11.0.0
    # via
    #   -r requirements.in
    #   matplotlib
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   django-qr-code
pydantic-core==2.23.4
    # via
    #   -r requirements.in
    #   pydantic
pymongo==4.10.1
    # via
    #   -r requirements.in
    #   mongoengine
pymysql==1.1.1
    # via -r requirements.in
pyparsing==3.2.0
    # via
    #   -r requirements.in
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   matplotlib
pytz==2024.2
    # via -r requirements.in
qrcode==8.0
    # via -r requirements.in
segno==1.6.1
    # via
    #   -r requirements.in
    #   django-qr-code
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via
    #   -r requirements.in
    #   django
whitenoise==6.8.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
